Separate and group config. tests
